Set up component skeleton
