+++ web/coincert/.gitignore
@@ -1,5 +1,6 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
+zoom_.mp4
 # dependencies
 /node_modules
 /.pnp
