+++ web/coincert/package.json
@@ -3,15 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@brightcove/videojs-flashls-source-handler": "^1.4.8",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-player": "^2.0.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
-    "video.js": "^7.7.5",
-    "videojs-flash": "^2.2.1"
+    "video.js": "^7.7.5"
   },
   "scripts": {
     "start": "react-scripts start",
