+++ web/coincert/src/Components/EventAbout.js
@@ -1,0 +1,1 @@
+// Display info about event, allow ticket purchase
