+++ web/coincert/src/Components/EventStreamerHost.js
@@ -1,0 +1,15 @@
+import React, {Component} from 'react';
+import { Player } from 'video-react';
+import '../App.css';
+
+class EventStreamerHost extends Component {
+  render() {
+    return (
+      <div>
+        <h1> TO DO <h2/>
+      </div>
+    )
+
+  }
+
+export default EventStreamerHost;
