+++ web/coincert/src/Components/MyEvents.js
@@ -1,0 +1,1 @@
+// Display events and profile information about the user currently logged in (use wallet as key in DB to store info?)
