+++ web/coincert/src/Components/SearchBar.js
@@ -1,0 +1,1 @@
+//SearchBar component. Maybe should be apart of the NabBar?
