+++ web/coincert/src/Components/SearchResults.js
@@ -1,0 +1,1 @@
+// Display results from the search. A list of events with links
