+++ web/coincert/src/Middleware/DBIntegration.js
@@ -1,0 +1,1 @@
+//Store API functions to access local database. If we want to store info off chain
