+++ web/coincert/src/Middleware/MetaMaskFunctions.js
@@ -1,0 +1,1 @@
+// Create Metamask wrapper functions here so they can be used across multiple components
